backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
psycopg2-binary>=2.9  # For PostgreSQL
gunicorn>=21.0  # For production
python-decouple>=3.8  # For environment variables